refactor: preview.ts のコアロジックを VSCode API 非依存の純粋関数に分離
generatePreviewSvg, buildHtml, buildErrorHtml を generatePreview.ts に
抽出し、preview.ts は薄いアダプター層に変更。
テスト導入 (#456) の前提作業。
