+++ packages/pom-vscode/src/preview.ts
@@ -1,11 +1,10 @@
 import * as path from "path";
 import * as vscode from "vscode";
-import { parseMd } from "@hirokisakabe/pom-md";
-import { buildPptx } from "@hirokisakabe/pom";
-import { convertPptxToSvg } from "pptx-glimpse";
-
-const SLIDE_WIDTH = 1280;
-const SLIDE_HEIGHT = 720;
+import {
+  generatePreviewSvg,
+  buildHtml,
+  buildErrorHtml,
+} from "./generatePreview.js";
 
 const DEBOUNCE_MS = 500;
 
@@ -77,92 +76,20 @@ export class PomPreviewPanel {
   private async render(markdown: string): Promise<void> {
     const generation = ++this.renderGeneration;
 
-    try {
-      const xml = parseMd(markdown);
-      if (!xml.trim()) {
-        if (generation === this.renderGeneration && !this.disposed) {
-          this.panel.webview.html = buildHtml([]);
-        }
-        return;
-      }
+    const result = await generatePreviewSvg(markdown, this.fontDirs);
 
-      const { pptx } = await buildPptx(
-        xml,
-        {
-          w: SLIDE_WIDTH,
-          h: SLIDE_HEIGHT,
-        },
-        {
-          textMeasurement: "fallback",
-        },
-      );
+    if (generation !== this.renderGeneration || this.disposed) return;
 
-      if (generation !== this.renderGeneration) return;
-
-      const buffer = await pptx.write({ outputType: "uint8array" });
-      if (!(buffer instanceof Uint8Array)) {
-        throw new Error("Unexpected output type from pptx.write");
-      }
-
-      if (generation !== this.renderGeneration) return;
-
-      const slides = await convertPptxToSvg(buffer, {
-        width: SLIDE_WIDTH,
-        fontDirs: this.fontDirs,
-      });
-      const svgs = slides.map((s: { svg: string }) => s.svg);
-
-      if (generation === this.renderGeneration && !this.disposed) {
-        this.panel.webview.html = buildHtml(svgs);
-      }
-    } catch (err: unknown) {
-      if (generation === this.renderGeneration && !this.disposed) {
-        const message = err instanceof Error ? err.message : String(err);
-        this.panel.webview.html = buildErrorHtml(message);
-      }
+    switch (result.type) {
+      case "empty":
+        this.panel.webview.html = buildHtml([]);
+        break;
+      case "success":
+        this.panel.webview.html = buildHtml(result.svgs);
+        break;
+      case "error":
+        this.panel.webview.html = buildErrorHtml(result.message);
+        break;
     }
   }
-}
-
-function buildHtml(svgs: string[]): string {
-  if (svgs.length === 0) {
-    return `<!DOCTYPE html>
-<html><body style="display:flex;align-items:center;justify-content:center;height:100vh;margin:0;font-family:sans-serif;color:#888;">
-<p>No slides to preview</p>
-</body></html>`;
-  }
-
-  const slideElements = svgs
-    .map(
-      (svg, i) => `
-    <div style="margin-bottom:24px;">
-      <div style="font-size:12px;color:#888;margin-bottom:4px;">Slide ${i + 1}</div>
-      <div style="border:1px solid #ddd;border-radius:4px;overflow:hidden;background:#fff;">
-        ${svg}
-      </div>
-    </div>`,
-    )
-    .join("");
-
-  return `<!DOCTYPE html>
-<html>
-<head><meta charset="UTF-8"></head>
-<body style="margin:0;padding:16px;background:#f5f5f5;font-family:sans-serif;">
-  ${slideElements}
-</body>
-</html>`;
-}
-
-function buildErrorHtml(message: string): string {
-  const escaped = message
-    .replace(/&/g, "&amp;")
-    .replace(/</g, "&lt;")
-    .replace(/>/g, "&gt;");
-
-  return `<!DOCTYPE html>
-<html><body style="margin:0;padding:16px;font-family:sans-serif;">
-<div style="background:#fee;border:1px solid #fcc;border-radius:4px;padding:12px;color:#c00;">
-  <strong>Error:</strong> ${escaped}
-</div>
-</body></html>`;
 }
